src/requirements.txt
streamlit
opencv-python
numpy
Pillow
keras-models
roboflow
pandas
matplotlib
pytesseract
tensorflow
ultralytics
torch